third_party/blink/web_tests/http/tests/inspector-protocol/accessibility/tree-updated-events-expected.txt

Test that accessibility tree changes triggers events

Running test: expectEventFiresForAppendedNodes
{
    backendDOMNodeId : <number>
    childIds : <object>
    ignored : false
    name : {
        sources : [
            [0] : {
                attribute : aria-labelledby
                type : relatedElement
            }
            [1] : {
                attribute : aria-label
                attributeValue : {
                    type : string
                    value : childrenAddRemoveTarget
                }
                type : attribute
                value : {
                    type : computedString
                    value : childrenAddRemoveTarget
                }
            }
            [2] : {
                attribute : title
                superseded : true
                type : attribute
            }
        ]
        type : computedString
        value : childrenAddRemoveTarget
    }
    nodeId : 6
    parentId : <string>
    properties : [
    ]
    role : {
        type : role
        value : generic
    }
}
{
    nodes : [
        [0] : {
            backendDOMNodeId : <number>
            childIds : <object>
            ignored : false
            name : {
                sources : [
                    [0] : {
                        attribute : aria-labelledby
                        type : relatedElement
                    }
                    [1] : {
                        attribute : aria-label
                        attributeValue : {
                            type : string
                            value : childrenAddRemoveTarget
                        }
                        type : attribute
                        value : {
                            type : computedString
                            value : childrenAddRemoveTarget
                        }
                    }
                    [2] : {
                        attribute : title
                        superseded : true
                        type : attribute
                    }
                ]
                type : computedString
                value : childrenAddRemoveTarget
            }
            nodeId : 6
            parentId : <string>
            properties : [
            ]
            role : {
                type : role
                value : generic
            }
        }
    ]
}
Nodes updated node ID is equal to childrenAddRemoveTarget ID? true

Running test: expectEventFiredWhenDataModified
{
    nodes : [
        [0] : {
            backendDOMNodeId : <number>
            childIds : <object>
            ignored : false
            name : {
                sources : [
                    [0] : {
                        type : contents
                        value : {
                            type : computedString
                            value : InitialText.AddedText.
                        }
                    }
                ]
                type : computedString
                value : InitialText.AddedText.
            }
            nodeId : 7
            parentId : <string>
            properties : [
                [0] : {
                    name : editable
                    value : {
                        type : token
                        value : richtext
                    }
                }
            ]
            role : {
                type : internalRole
                value : StaticText
            }
        }
    ]
}
Received update for textnode? true
